+++ core/src/zeit/content/cp/blocks/tests/test_doctests.py
@@ -6,6 +6,5 @@ def test_suite():
     return zeit.content.cp.testing.FunctionalDocFileSuite(
         'cpextra.txt',
         'teaser.txt',
-        'xml.txt',
         package=zeit.content.cp.blocks,
     )