+++ core/src/zeit/content/cp/browser/blocks/audio.py
@@ -1,0 +1,6 @@
+import zeit.content.cp.browser.blocks.block
+import zeit.content.cp.interfaces
+
+
+class EditProperties(zeit.content.cp.browser.blocks.block.EditCommon):
+    pass